client/requirements.txt
grpcio==1.51.1 
grpcio-tools==1.51.1 
protobuf==3.19
pyaudio
matplotlib
torch
torchvision 
torchaudio